Riak 2.0pre4 with strong consistency (Work In Progress)
K/V operations to buckets named "c~bucket", "c~testing", etc
are consistent. Operations to other buckets are normal Riak
operations.

Consistent data cannot generate siblings, and uses optimistic
concurrency control. A get/modify/put cycle against a consistent
object will fail if someone modified the object concurrently.

Operations to individual keys are linearizable/atomic. If a
write succeeds, the next read will reflect that write/a later
write.

If a write fails, it may have been a partial write. The value
is unknown. However, the next successful read resolves the
ambiguity. The read will either show the old value or the value
of the failed write. Once resolved, the value won't change. In
other words, if a read shows the old value, you're guaranteed
to never see the failed-to-write value appear in the future.

Internally, there are three types of puts: put_once, modify,
and overwrite. put_once and modify are safe/linearizable operations.
Both fail if concurrent operations break their preconditions.
Overwrite is an unchecked put and is unsafe.

NOTE: In this work-in-progress branch, client puts that are missing
vector clocks correspond to overwrites which can violate linearizablity.
This will soon change to map empty-vclock operations to put_once
instead, with a new client option to explicitly request overwrite
semantics.

Minor notes:

-- Consistent data does not currently support secondary indexes.
-- Consistent operations are not yet reflected in Riak's stats.

THIS IS A WORK IN PROGRESS

This code has been tested, but is still pending code review. It
will be reviewed/cleaned-up/merged into Riak's develop branch in
the near future.